Moved MLIROp, transformation pipeline and evaluation tools to a separate files mlir_op.hpp/cpp
